unmocking prove_recursively and creating scheduler bin
