+++ Cargo.toml
@@ -4,7 +4,7 @@ resolver = "2"
 members = [
   "bin/dojo-language-server",
   "bin/katana",
-  "bin/saya",
+  "bin/saya", "bin/scheduler",
   "bin/sozo",
   "bin/torii",
   "crates/benches",
