+++ bin/scheduler/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "scheduler"
+edition.workspace = true
+license.workspace = true
+license-file.workspace = true
+repository.workspace = true
+version.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+saya-core.workspace = true
+katana-primitives.workspace = true
+tokio.workspace = true
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
