+++ crates/saya/core/src/prover/mod.rs
@@ -7,7 +7,7 @@ use anyhow::bail;
 use async_trait::async_trait;
 
 mod program_input;
-mod scheduler;
+pub mod scheduler;
 mod serializer;
 pub mod state_diff;
 mod stone_image;
@@ -26,7 +26,7 @@ pub enum ProverIdentifier {
     Stone,
     Sharp,
     Platinum,
-    #[cfg(test)]
+    //#[cfg(test)]
     Dummy,
 }
 
@@ -35,7 +35,7 @@ pub async fn prove(input: String, prover: ProverIdentifier) -> anyhow::Result<St
         ProverIdentifier::Sharp => todo!(),
         ProverIdentifier::Stone => prove_stone(input).await,
         ProverIdentifier::Platinum => todo!(),
-        #[cfg(test)]
+        //#[cfg(test)]
         ProverIdentifier::Dummy => {
             tokio::time::sleep(std::time::Duration::from_secs(2)).await;
             Ok("dummy ok".to_string())
